fix(cli): split send-prompt out of api/sessions.go (over 1000-line limit), inventory master's project-context
CI failed the file-length lint: api/sessions.go was 1017 lines in the PR merge —
master's #1893 (project-context contract) grew it to 963 and this PR's --prompt
alias tipped it over 1000. Per repo policy the fix is to SPLIT, never grandfather.

Moved the whole send-prompt unit — the flag vars, the argument-shape helpers
(sendPromptArgCount/sendPromptUsage/resolveSendPrompt added here), the command,
validateSendPromptFlags, and runBroadcast — into api/sessions_sendprompt.go. It
is a cohesive group, same package so every cross-reference still resolves, and it
drops api/sessions.go to 658 lines: 300+ of headroom against master's continued
growth, not a hairline pass. The --prompt alias behaves identically after the
move (re-driven through the built binary).

Rebased onto current master (was 4 further behind) and inventoried the surfaces
those commits added, so the audit passes honestly:
  - task.repo_id (#1893): daemon-resolved at bind time, no client sets it —
    declared `ok` for every surface, like created_at.
  - expect / expect.enforce / expect.project_path (#1891 ProjectExpectation): a
    CLI-side compare-and-swap guard (enforceTaskScope authorizes a mutation
    against the project the CLI resolved via --repo). The CLI sets it via a
    literal, so the derivation credits it; the TUI/web operate within an
    already-selected project and set no expectation — declared `ok` there, a
    safety mechanism, not a user option.
  - af tasks list --all: list across every project — mapped to task.list.

Every rebase this session, the audit caught exactly what master had changed —
which is the whole point of the PR, demonstrated against a live, fast-moving
master rather than argued.

Container suite green (33 packages incl. api, daemon 156s, parity). Coverage
denominator grew (more surface read), SKIPPED: none.
